+++ NEWS.md
@@ -1,5 +1,6 @@
 # worldfootballR 0.0.0.9000
 
 * First commit of package
-* Added function to get the match results of a given tier-1 league season
-* Added function to extract match URLs for a given tier-1 league season
+* Added function to get the match results of a given tier-1 league season `get_match_results`
+* Added function to extract match URLs for a given tier-1 league season `get_match_urls`
+* Added function to get team season statistics `get_season_team_stats`